Experiment another way to get all the files needed to be included into a plt.
This is still highly experimental and some features are broken now. From what I could see it's a hard path this one, because we are currently getting only the modules that are called directly from another module, and also it's hard to analyze correctly all the modules of an application without raising too much the time.